fix: redact persisted connector arguments
