+++ changelogs/v2.5.22.md
@@ -11,3 +11,4 @@
 - Connector tool cards are now collapsed by default and expandable on desktop and mobile.
 - Connector tool details now hide internal connector IDs and pretty-print structured results.
 - Expanded connector results now show up to three discovered links with a Show more control.
+- Redact sensitive connector arguments before they enter persisted desktop and mobile chat history.
