+++ mobile/changelogs/v0.1.5.md
@@ -5,3 +5,4 @@
 - Connector tool cards are now collapsed by default and expandable to show the called tool and result.
 - Connector tool details now hide internal connector IDs and pretty-print structured results.
 - Expanded connector results now show up to three discovered links with a Show more control.
+- Redact sensitive connector arguments before they enter persisted mobile chat history.
